Change filters back to string only

+++ lib/Filter/Rule/Float.php
@@ -12,6 +12,6 @@ class Float extends Rule
 {
     protected function _filter($value)
     {
-        return (float) filter_var($value, FILTER_SANITIZE_NUMBER_FLOAT, FILTER_FLAG_ALLOW_FRACTION);
+        return filter_var($value, FILTER_SANITIZE_NUMBER_FLOAT, FILTER_FLAG_ALLOW_FRACTION);
     }
 }
